feat: create LocalStorageFileSystem/ORM
