+++ src/LocalStorageFileSystem/ORM/DB.ts
@@ -1,0 +1,22 @@
+import ORM from "./ORM";
+
+class DB extends ORM {
+
+    constructor() {
+
+        super()
+
+    }
+    /**
+     * only for development
+     */
+    public drop() {
+
+        this.db.drop()
+        console.log('dropped!')
+
+    }
+    
+}
+
+export default DB
