+++ src/LocalStorageFileSystem/ORM/index.ts
@@ -1,0 +1,9 @@
+import DB from "./DB";
+import FileORM from "./FileORM";
+import DirectoryORM from "./DirectoryORM";
+
+export {
+    DB,
+    FileORM as File,
+    DirectoryORM as Directory
+}
