--- src/Storage/index.ts
@@ -1,9 +1,0 @@
-import Directory, { DirectoryNullable } from "./directory";
-import File, { FileNullable } from "./file";
-
-export {
-    Directory,
-    DirectoryNullable,
-    File,
-    FileNullable
-}
